// src/app/features/lend-unrealdn-marketvalue/models/unreal-detail.model.ts
export interface UnrealDetail {
  tdate: string;
  stock: string;
  ttypename: string;
  dseq: string;
  qty: string;
  bqty: string;
  price: string;
  mamt: string;
  fee: string;
  tax: string;
  divamt: string;
  interest: string;
  cost: string;
  wtype: string;
  seqno: string;
  nowamt: string;
  unreal: string;
  currnm: string;
  ioflag: string;
}