// src/app/features/lend-unrealdn-marketvalue/models/unreal-sum.model.ts
export interface UnrealSum {
  no: string;
  ttypename: string;
  stock: string;
  stocknm: string;
  real_qty: string;
  qty: string;
  bqty: string;
  sqty: string;
  cost: string;
  avgprice: string;
  nowamt: string;
  unreal: string;
  urratio: string;
  lastprice: string;
  interest: string;
  divamt: string;
  AD: string;
  ADR: string;
  currnm: string;
}